ktime: Get rid of the union
ktime is a union because the initial implementation stored the time in
scalar nanoseconds on 64 bit machine and in a endianess optimized timespec
variant for 32bit machines. The Y2038 cleanup removed the timespec variant
and switched everything to scalar nanoseconds. The union remained, but
become completely pointless.

Get rid of the union and just keep ktime_t as simple typedef of type s64.

The conversion was done with coccinelle and some manual mopping up.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Change-Id: I2897083798688e2681d11ba328e367591b65b167
Cc: Peter Zijlstra <peterz@infradead.org>
Signed-off-by: kdrag0n <dragon@khronodragon.com>
Signed-off-by: DennySPB <dennyspb@gmail.com>

Conflicts:
       include/linux/futex.h

Signed-off-by: TogoFire <italomellopereira@gmail.com>
Signed-off-by: Marco Zanin <mrczn.bb@gmail.com>